# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Система оценок (Gradebook API)",
        "description": "API для управления студентами, предметами и их оценками. Использованы Laravel Resources и hasManyThrough.",
        "contact": {
            "email": "artemtaiti@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel API Server"
        }
    ],
    "paths": {
...
        "/subjects": {
            "get": {
                "tags": [
                    "Предметы"
                ],
                "summary": "Получить список всех предметов",
                "operationId": "2a7ea80416f2d2d5be939b5287b03bc0",
                "responses": {
                    "200": {
                        "description": "Успешный ответ. Возвращает коллекцию предметов с их связями.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SubjectResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Предметы"
                ],
                "summary": "Создать новый предмет",
                "operationId": "1f14c36996b396b8c8e42db001a0e017",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "lecturer_name"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "История"
                                    },
                                    "lecturer_name": {
                                        "type": "string",
                                        "example": "Петров П.П."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Предмет успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SubjectResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/subjects/{subject}": {
            "get": {
                "tags": [
                    "Предметы"
                ],
                "summary": "Получить предмет по ID",
                "description": "Включает в себя Eager Loading связей grades и students.",
                "operationId": "08787fe368cbd43448d0c4e1c5c956d5",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "ID предмета",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SubjectResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Предмет не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Предметы"
                ],
                "summary": "Обновить данные предмета",
                "operationId": "ab57b575049f7e7b92e6ccc7ac6fb12b",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "ID предмета",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "lecturer_name"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "История (Обновлено)"
                                    },
                                    "lecturer_name": {
                                        "type": "string",
                                        "example": "Петров П.П."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SubjectResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Предмет не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Предметы"
                ],
                "summary": "Удалить предмет",
                "operationId": "c6dece96323c0586fd57b7b721f125fa",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "ID предмета",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Предмет успешно удален (Нет содержимого)"
                    },
                    "404": {
                        "description": "Предмет не найден"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SubjectResource": {
                "title": "Subject Resource (Output)",
                "description": "Предмет, возвращаемый API, включая связанные оценки и студентов.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "title": {
                        "description": "Название предмета",
                        "type": "string",
                        "example": "Высшая математика"
                    },
                    "lecturer_name": {
                        "description": "Имя лектора",
                        "type": "string",
                        "example": "Иванов И.И."
                    },
                    "grades": {
                        "description": "Список оценок по этому предмету (включая студента)",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 5
                                },
                                "score": {
                                    "type": "integer",
                                    "example": 5
                                },
                                "exam_date": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2025-11-20"
                                },
                                "student_id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            },
                            "type": "object"
                        }
                    },
                    "students": {
                        "description": "Список студентов, получивших оценки по этому предмету (из hasManyThrough).",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "first_name": {
                                    "type": "string",
                                    "example": "Иван"
                                },
                                "last_name": {
                                    "type": "string",
                                    "example": "Иванов"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}